+++ src/containers/Coins.js
@@ -7,6 +7,7 @@ import CategoryMenu from '../components/CategoryMenu';
 import LoadingIcon from '../components/Loader';
 import CoinCard from '../components/CoinCard';
 import getCoins from '../actions/index';
+import sort from '../modules/sort';
 
 class CoinsContainer extends Component {
   constructor(props) {
@@ -21,35 +22,10 @@ class CoinsContainer extends Component {
   render() {
     const { coin } = this.props.state;
     const { activeItem } = this.props.state.filter;
-    let sortedCoins;
     if (coin.length <= 0) {
       return (<LoadingIcon />);
     }
-
-    const descRank = () => {
-      const sortedCoins = [];
-      for (let i = coin[0].length - 1; i >= 0; i -= 1) {
-        sortedCoins.push(coin[0][i]);
-      }
-      return sortedCoins;
-    };
-
-    const ascRank = () => coin[0];
-
-    const filterByPrice = () => {
-      sortedCoins = [...coin[0]];
-      sortedCoins.sort((a, b) => a.price - b.price);
-      return sortedCoins;
-    };
-
-    if (activeItem === 'ASC_RANK') {
-      sortedCoins = ascRank(); // default
-    } else if (activeItem === 'PRICE') {
-      sortedCoins = filterByPrice();
-    } else {
-      sortedCoins = descRank();//
-    }
-    //  this.props.state.filer.activeItem = "ASC_RANK" NOW FILTER the components =>
+    const sortedCoins = sort(coin, activeItem);
     return (
       <Container>
         <CategoryMenu />
